backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
fastapi==0.136.0
google-genai==1.73.1
h11==0.16.0
idna==3.13
networkx==3.6.1
numpy==2.3.5
pydantic==2.13.3
pydantic-settings==2.11.0
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
scikit-learn==1.8.0
starlette==1.0.0
tensorflow==2.20.0; python_version < "3.14"
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
websockets==16.0
xgboost==3.1.2
